Fix MemAlloc receiving an incorrect size

+++ MonoMod.RuntimeDetour/Platforms/Native/DetourNativeARMPlatform.cs
@@ -187,8 +187,8 @@ namespace MonoMod.RuntimeDetour.Platforms {
             // no-op.
         }
 
-        public IntPtr MemAlloc(int size) {
-            return Marshal.AllocHGlobal(size);
+        public IntPtr MemAlloc(uint size) {
+            return Marshal.AllocHGlobal((int) size);
         }
 
         public void MemFree(IntPtr ptr) {
