+++ MonoMod.RuntimeDetour/Platforms/Native/DetourNativeX86Platform.cs
@@ -116,8 +116,8 @@ namespace MonoMod.RuntimeDetour.Platforms {
             // no-op.
         }
 
-        public IntPtr MemAlloc(int size) {
-            return Marshal.AllocHGlobal(size);
+        public IntPtr MemAlloc(uint size) {
+            return Marshal.AllocHGlobal((int) size);
         }
 
         public void MemFree(IntPtr ptr) {
